+++ tools/tools.go
@@ -5,5 +5,6 @@ package tools
 
 import (
 	_ "golang.org/x/lint/golint"
+	_ "golang.org/x/tools/cmd/stringer"
 	_ "honnef.co/go/tools/cmd/staticcheck"
 )